# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreatePremiseDto:
    properties:
      address:
        maxLength: 255
        minLength: 2
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      parent_premise_id:
        type: string
    required:
    - address
    - name
    type: object
...
  dto.UpdatePremiseDto:
    properties:
      address:
        maxLength: 255
        minLength: 2
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - address
    - name
    type: object
  dto.UpdatePremiseUserDto:
    properties:
      added_users:
        items:
          type: string
        type: array
      removed_users:
        items:
          type: string
        type: array
    type: object
  errors.ErrorDetail:
    properties:
      details: {}
      message:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      request_id:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
  errors.ErrorType:
    enum:
    - VALIDATION_ERROR
    - NOT_FOUND
    - UNAUTHORIZED
    - FORBIDDEN
    - BAD_REQUEST
    - CONFLICT
    - INTERNAL_ERROR
    - DATABASE_ERROR
    - EXTERNAL_SERVICE_ERROR
    - TIMEOUT_ERROR
    type: string
    x-enum-varnames:
    - ErrorTypeValidation
    - ErrorTypeNotFound
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeInternal
    - ErrorTypeDatabase
    - ErrorTypeExternal
    - ErrorTypeTimeout
...
  models.Premise:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      parent_premise:
        $ref: '#/definitions/models.Premise'
      parent_premise_id:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
  types.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
    type: object
  types.PremiseListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Premise'
        type: array
      pagination:
        $ref: '#/definitions/types.Pagination'
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Smart City System (SCS) Operator API for managing premises, alarms,
    incidents, guidance templates, and guards
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SCS Operator API
  version: "1.0"
paths:
...
  /premises:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all premises
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PremiseListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get premises with pagination
      tags:
      - premises
    post:
      consumes:
      - application/json
      description: Create a new premise with the provided information
      parameters:
      - description: Premise creation data
        in: body
        name: premise
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePremiseDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Premise'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new premise
      tags:
      - premises
  /premises/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific premise by its ID
      parameters:
      - description: Premise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Premise'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get premise by ID
      tags:
      - premises
    patch:
      consumes:
      - application/json
      description: Update an existing premise with new information
      parameters:
      - description: Premise ID
        in: path
        name: id
        required: true
        type: string
      - description: Premise update data
        in: body
        name: premise
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePremiseDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Premise'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update premise
      tags:
      - premises
  /premises/{id}/assign-users:
    post:
      consumes:
      - application/json
      description: Add or remove users from a premise
      parameters:
      - description: Premise ID
        in: path
        name: id
        required: true
        type: string
      - description: User assignment data
        in: body
        name: users
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePremiseUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Assign users to premise
      tags:
      - premises
  /premises/{id}/users:
    get:
      consumes:
      - application/json
      description: Get all users assigned to a specific premise
      parameters:
      - description: Premise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get users assigned to premise
      tags:
      - premises
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"